+++ apps/blog/src/app/components/pages/home/components/PostCardsPage.tsx
@@ -1,6 +1,6 @@
 import type { PostCardRo } from "~/api/models";
-import { PostCards } from "~/app/components/postCards";
-import { PostsProvider } from "~/app/components/postCards/components/Context";
+import { PostCards } from "~/app/components/misc/postCards";
+import { PostsProvider } from "~/app/components/misc/postCards/components/Context";
 
 export function PostCardsPage({ posts }: { posts: PostCardRo[] }) {
   return (
